lincoln_weather {ggridges}	R Documentation

Weather in Lincoln, Nebraska in 2016.

Description

A dataset containing weather information from Lincoln, Nebraska, from 2016. Originally downloaded from Weather Underground by Austin Wehrwein, http://austinwehrwein.com/. The variables are listed below. Most are self-explanatory. Max, mean, and min measurements are calculated relative to the specific day of measurement.

Usage


lincoln_weather

Format

A tibble with 366 rows and 24 variables:

CST

Day of the measurement

Max Temperature [F]

Mean Temperature [F]

Min Temperature [F]

Max Dew Point [F]

Mean Dew Point [F]

Min Dewpoint [F]

Max Humidity

Mean Humidity

Min Humidity

Max Sea Level Pressure [In]

Mean Sea Level Pressure [In]

Min Sea Level Pressure [In]

Max Visibility [Miles]

Mean Visibility [Miles]

Min Visibility [Miles]

Max Wind Speed [MPH]

Mean Wind Speed[MPH]

Max Gust Speed [MPH]

Precipitation [In]

CloudCover

Events

Specific weather events, such as rain, snow, or fog

WindDir [Degrees]

Month

The month in which the measurement was taken

[Package ggridges version 0.5.3 Index]